Изменение бд на MsSQL

+++ src/main/groovy/factory/first/may/backend/request_models/request/SellRequest.groovy
@@ -3,7 +3,7 @@ package factory.first.may.backend.request_models.request
 class SellRequest {
     Long idSell
 
-    Long sum
+    double sum
 
     Date dateSell
